+++ src/fmalloc.cpp
@@ -7,4 +7,5 @@
 #include "fmalloc_ooc.inc"
 #include "fmalloc_tensor.inc"
 #include "fmalloc_alp.inc"
+#include "fmalloc_insitu.inc"
 #include "fmalloc_api.inc"